refactor(routes): extract shared helpers from agents.ts to agents-helpers.ts
Moves AgentSummary/AgentDetail types, getAgentSummary, getAgentDetail,
listAgentSummaries, matchChannelRoute, VALID_PROVIDERS, remote caches,
and related utilities to a dedicated helpers module.

agents.ts re-exports validateDiscordChannelId for backward compatibility
so existing tests continue importing from the same path.

agents.ts: 1921 -> 1709 lines | agents-helpers.ts: 277 lines new
